+++ test/td_dd_web/controllers/xlsx_controller_test.exs
@@ -1,5 +1,5 @@
 defmodule TdDdWeb.DataStructures.XLSXControllerTest do
-  use Oban.Testing, repo: TdDd.Repo, prefix: "private"
+  use Oban.Testing, repo: TdDd.Repo, prefix: Application.get_env(:td_dd, Oban)[:prefix]
   use TdDdWeb.ConnCase
 
   import Mox